Promote ov002/da1up_c: 36 shards into one TU, under the ROM's class name
Folds the contiguous linker run 0x020aee40..0x020b0530 -- 36 functions, ROM
ordinals 0..35 -- into src/actors/da1up_c.cpp, and moves the class off its
coined name at the same time.

THE NAME IS THE CARTRIDGE'S. ov002 file offset 0x5ad10 == address 0x02108370
holds `7da1up_c\0`, the length-prefixed Itanium type-name string, and
_ZTI7da1up_c at 0x0210837c is the matching __si_class_type_info whose +8 word
reaches _ZTI12dEnemyBase_c at 0x021081c0. `OneUpMushroom` is absent from every
executable image in every encoding tested. Nine mangled member symbols, the
header and the _ZTV alias all move in this one commit, because dsd resolves
delinked .data PMF records by symbol NAME and a half-landed rename links them
as 0x00000000.

THE RENAME IS PROVEN BYTE-NEUTRAL. `linkcheck --baseline` was re-run on the
renamed tree before any other change and produced modules PASS and the same ROM
sha256 d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8 as the
pre-rename baseline control.

VERIFY: 36/36 MATCH, objisolate clean, reloc-destinations clean, all 36 in the
expected ROM-ascending section order, 13-row compiler_only_output policy
accepted -> TEXT-VERIFIED.
LINKCHECK (pre-promotion, the only usable kind): SCRATCH-LINK-VERIFIED. Module
fidelity 106/106 exact; the full ROM builds to the sha256 above, IDENTICAL to
the strict stock control. `dsd check symbols --fail` still FAILS on the same 9
pre-existing errors the baseline control has (0 new) -- not green, and not
claimed as such.

METHODS: 9 of the 36 symbols are real da1up_c:: methods, written as 8
definitions (ordinals 0 and 1 are two shards of the one destructor), and all 9
are this class's own vtable slots -- 16/17 the destructor pair, 19
OnTurnIntoEgg, 18 OnYoshiTryEat, 3 CleanupResources, 12 OnPendingDestroy, 9
Render, 6 Behavior, 0 InitResources. Five of those were declared non-virtual in
the old header, which the cartridge's own 31-slot table at 0x021083c8
contradicts; they are virtual now. The other 27 hit the NAMING wall and stay
free functions: 14 have ROM-proven member-ness and a ROM-proven dispatch index
(the 14 {fn, 0} PMF descriptors at 0x02108300..0x02108370 that
__sinit_ov002_02100adc copies to 0x0210dc00) but no name evidence, and 13 have
no member-ness evidence at all. No codegen wall and no scope wall was hit.

FORM: out-of-line destructor declared first (so this TU owns the key function
and emits the chain's vtable/RTTI as vague-linkage passengers), ROM-ascending
source order under `#pragma defer_codegen off`. That reproduces the cartridge's
D1 at 0x020aee40 below D0 at 0x020aee88 with a homeless trailing D2, and leaves
only 3 upward intra-TU calls, so each member keeps its own recovered spelling.
decl_common.h is deliberately not included: it declares 11 of these 36 symbols
and 4 disagree with the byte-matched body.

RATCHET: 5 banked rows swap 1:1 from the deleted shard paths to
src/actors/da1up_c.cpp#<member>; baseline 2696 -> 2696.
