+++ config/arm9/overlays/ov002/delinks.txt
@@ -140,149 +140,9 @@ src/_ZN11dCapEnemy_cD0Ev.cpp:
     complete
     .text start:0x020aedf4 end:0x020aee40
 
-src/_ZN13OneUpMushroomD1Ev.cpp:
+src/actors/da1up_c.cpp:
     complete
-    .text start:0x020aee40 end:0x020aee88
-
-src/_ZN13OneUpMushroomD0Ev.cpp:
-    complete
-    .text start:0x020aee88 end:0x020aeee4
-
-src/func_ov002_020aeee4.cpp:
-    complete
-    .text start:0x020aeee4 end:0x020aefa4
-
-src/func_ov002_020aefa4.c:
-    complete
-    .text start:0x020aefa4 end:0x020aefb8
-
-src/func_ov002_020aefb8.cpp:
-    complete
-    .text start:0x020aefb8 end:0x020af0c0
-
-src/func_ov002_020af0c0.c:
-    complete
-    .text start:0x020af0c0 end:0x020af1dc
-
-src/func_ov002_020af1dc.c:
-    complete
-    .text start:0x020af1dc end:0x020af218
-
-src/func_ov002_020af218.c:
-    complete
-    .text start:0x020af218 end:0x020af248
-
-src/func_ov002_020af248.c:
-    complete
-    .text start:0x020af248 end:0x020af2b0
-
-src/_ZN13OneUpMushroom13OnTurnIntoEggER6Player.cpp:
-    complete
-    .text start:0x020af2b0 end:0x020af3a0
-
-src/_ZN13OneUpMushroom13OnYoshiTryEatEv.cpp:
-    complete
-    .text start:0x020af3a0 end:0x020af3a8
-
-src/func_ov002_020af3a8.c:
-    complete
-    .text start:0x020af3a8 end:0x020af474
-
-src/func_ov002_020af474.c:
-    complete
-    .text start:0x020af474 end:0x020af4ec
-
-src/func_ov002_020af4ec.cpp:
-    complete
-    .text start:0x020af4ec end:0x020af684
-
-src/func_ov002_020af684.cpp:
-    complete
-    .text start:0x020af684 end:0x020af724
-
-src/func_ov002_020af724.c:
-    complete
-    .text start:0x020af724 end:0x020af7cc
-
-src/func_ov002_020af7cc.c:
-    complete
-    .text start:0x020af7cc end:0x020af838
-
-src/func_ov002_020af838.c:
-    complete
-    .text start:0x020af838 end:0x020af908
-
-src/func_ov002_020af908.c:
-    complete
-    .text start:0x020af908 end:0x020af924
-
-src/func_ov002_020af924.c:
-    complete
-    .text start:0x020af924 end:0x020af950
-
-src/func_ov002_020af950.c:
-    complete
-    .text start:0x020af950 end:0x020afa50
-
-src/func_ov002_020afa50.c:
-    complete
-    .text start:0x020afa50 end:0x020afa6c
-
-src/func_ov002_020afa6c.c:
-    complete
-    .text start:0x020afa6c end:0x020afa98
-
-src/func_ov002_020afa98.c:
-    complete
-    .text start:0x020afa98 end:0x020afbb4
-
-src/func_ov002_020afbb4.c:
-    complete
-    .text start:0x020afbb4 end:0x020afc44
-
-src/func_ov002_020afc44.c:
-    complete
-    .text start:0x020afc44 end:0x020afc68
-
-src/func_ov002_020afc68.cpp:
-    complete
-    .text start:0x020afc68 end:0x020afd10
-
-src/func_ov002_020afd10.c:
-    complete
-    .text start:0x020afd10 end:0x020afde4
-
-src/func_ov002_020afde4.cpp:
-    complete
-    .text start:0x020afde4 end:0x020afe4c
-
-src/func_ov002_020afe4c.cpp:
-    complete
-    .text start:0x020afe4c end:0x020aff10
-
-src/func_ov002_020aff10.c:
-    complete
-    .text start:0x020aff10 end:0x020affe8
-
-src/_ZN13OneUpMushroom16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x020affe8 end:0x020b006c
-
-src/_ZN13OneUpMushroom16OnPendingDestroyEv.cpp:
-    complete
-    .text start:0x020b006c end:0x020b0070
-
-src/_ZN13OneUpMushroom6RenderEv.cpp:
-    complete
-    .text start:0x020b0070 end:0x020b00e8
-
-src/_ZN13OneUpMushroom8BehaviorEv.cpp:
-    complete
-    .text start:0x020b00e8 end:0x020b01c0
-
-src/_ZN13OneUpMushroom13InitResourcesEv.cpp:
-    complete
-    .text start:0x020b01c0 end:0x020b0530
+    .text start:0x020aee40 end:0x020b0530
 
 src/d_a_1up_scaleup_kinoko.c:
     complete
