+++ config/arm9/overlays/ov002/symbols.txt
@@ -32,8 +32,8 @@ _ZN12dEnemyBase_cD1Ev kind:function(arm,size=0x24) addr:0x020aed74
 _ZN12dEnemyBase_cC2Ev kind:function(arm,size=0x24) addr:0x020aed98
 _ZN11dCapEnemy_cD2Ev kind:function(arm,size=0x38) addr:0x020aedbc
 _ZN11dCapEnemy_cD0Ev kind:function(arm,size=0x4c) addr:0x020aedf4
-_ZN13OneUpMushroomD1Ev kind:function(arm,size=0x48) addr:0x020aee40
-_ZN13OneUpMushroomD0Ev kind:function(arm,size=0x5c) addr:0x020aee88
+_ZN7da1up_cD1Ev kind:function(arm,size=0x48) addr:0x020aee40
+_ZN7da1up_cD0Ev kind:function(arm,size=0x5c) addr:0x020aee88
 func_ov002_020aeee4 kind:function(arm,size=0xc0) addr:0x020aeee4
 func_ov002_020aefa4 kind:function(arm,size=0x14) addr:0x020aefa4
 func_ov002_020aefb8 kind:function(arm,size=0x108) addr:0x020aefb8
@@ -41,8 +41,8 @@ func_ov002_020af0c0 kind:function(arm,size=0x11c) addr:0x020af0c0
 func_ov002_020af1dc kind:function(arm,size=0x3c) addr:0x020af1dc
 func_ov002_020af218 kind:function(arm,size=0x30) addr:0x020af218
 func_ov002_020af248 kind:function(arm,size=0x68) addr:0x020af248
-_ZN13OneUpMushroom13OnTurnIntoEggER6Player kind:function(arm,size=0xf0) addr:0x020af2b0
-_ZN13OneUpMushroom13OnYoshiTryEatEv kind:function(arm,size=0x8) addr:0x020af3a0
+_ZN7da1up_c13OnTurnIntoEggER6Player kind:function(arm,size=0xf0) addr:0x020af2b0
+_ZN7da1up_c13OnYoshiTryEatEv kind:function(arm,size=0x8) addr:0x020af3a0
 func_ov002_020af3a8 kind:function(arm,size=0xcc) addr:0x020af3a8
 func_ov002_020af474 kind:function(arm,size=0x78) addr:0x020af474
 func_ov002_020af4ec kind:function(arm,size=0x198) addr:0x020af4ec
@@ -63,11 +63,11 @@ func_ov002_020afd10 kind:function(arm,size=0xd4) addr:0x020afd10
 func_ov002_020afde4 kind:function(arm,size=0x68) addr:0x020afde4
 func_ov002_020afe4c kind:function(arm,size=0xc4) addr:0x020afe4c
 func_ov002_020aff10 kind:function(arm,size=0xd8) addr:0x020aff10
-_ZN13OneUpMushroom16CleanupResourcesEv kind:function(arm,size=0x84) addr:0x020affe8
-_ZN13OneUpMushroom16OnPendingDestroyEv kind:function(arm,size=0x4) addr:0x020b006c
-_ZN13OneUpMushroom6RenderEv kind:function(arm,size=0x78) addr:0x020b0070
-_ZN13OneUpMushroom8BehaviorEv kind:function(arm,size=0xd8) addr:0x020b00e8
-_ZN13OneUpMushroom13InitResourcesEv kind:function(arm,size=0x370) addr:0x020b01c0
+_ZN7da1up_c16CleanupResourcesEv kind:function(arm,size=0x84) addr:0x020affe8
+_ZN7da1up_c16OnPendingDestroyEv kind:function(arm,size=0x4) addr:0x020b006c
+_ZN7da1up_c6RenderEv kind:function(arm,size=0x78) addr:0x020b0070
+_ZN7da1up_c8BehaviorEv kind:function(arm,size=0xd8) addr:0x020b00e8
+_ZN7da1up_c13InitResourcesEv kind:function(arm,size=0x370) addr:0x020b01c0
 da1up_c_classInit_SCALEUP_KINOKO kind:function(arm,size=0x50) addr:0x020b0530
 da1up_c_classInit_ONEUPKINOKO kind:function(arm,size=0x50) addr:0x020b0580
 _ZN7daBar_cD1Ev kind:function(arm,size=0x30) addr:0x020b05d0
@@ -1690,7 +1690,6 @@ _ZTS7da1up_c kind:data(any) addr:0x02108370
 _ZTI7da1up_c kind:data(any) addr:0x0210837c
 g_profile_ONEUPKINOKO kind:data(any) addr:0x02108388
 g_profile_SCALEUP_KINOKO kind:data(any) addr:0x021083a4
-_ZTV13OneUpMushroom kind:data(any) addr:0x021083c8
 _ZTV7da1up_c kind:data(any) addr:0x021083c8
 _ZTS7daBar_c kind:data(any) addr:0x02108444
 _ZTI7daBar_c kind:data(any) addr:0x02108450
