+++ config/converted-backslide-exceptions.jsonl
@@ -395,3 +395,8 @@
 {"path": "src/_ZN10Scuttlebug16OnAimedAtWithEggEv.cpp", "reason": "ov071/Scuttlebug TU promotion, stage 3b: these five rows are a path rename, not a regression. Each of the five whole-file identities was absorbed verbatim into src/actors/Scuttlebug.cpp by the promotion and re-enters the set as a member identity (src/actors/Scuttlebug.cpp#Scuttlebug_SetState, #_ZN10Scuttlebug13OnYoshiTryEatEv, #_ZN10Scuttlebug16OnAimedAtWithEggEv, #_ZN10Scuttlebug16OnPendingDestroyEv, #_ZN10Scuttlebug6RenderEv), so the banked count is unchanged at 2686 -- 5 gained, 5 lost, same five units of code. classify_missing() reports them as MOVED with a member regression because OTHER members of the merged TU fail no_mangled_refs: a reconstructed TU must spell the vague-linkage symbols _ZN7fBase_cnwEj, _ZN8dActor_cC2Ev and _ZN8dActor_cD2Ev directly or the 0x0211f0a4..0x02120668 range will not link, and the 24 still-unnamed func_ov071_* members reach fields through this-relative arithmetic. Byte-match outranks readability: 35/35 MATCH, objisolate clean, reloc destinations clean, emission order ROM-ascending, linkcheck SCRATCH-LINK-VERIFIED with tuRange differingBytes 0, and rombuild 106/106 modules exact against the stock ROM."}
 {"path": "src/_ZN10Scuttlebug16OnPendingDestroyEv.cpp", "reason": "ov071/Scuttlebug TU promotion, stage 3b: these five rows are a path rename, not a regression. Each of the five whole-file identities was absorbed verbatim into src/actors/Scuttlebug.cpp by the promotion and re-enters the set as a member identity (src/actors/Scuttlebug.cpp#Scuttlebug_SetState, #_ZN10Scuttlebug13OnYoshiTryEatEv, #_ZN10Scuttlebug16OnAimedAtWithEggEv, #_ZN10Scuttlebug16OnPendingDestroyEv, #_ZN10Scuttlebug6RenderEv), so the banked count is unchanged at 2686 -- 5 gained, 5 lost, same five units of code. classify_missing() reports them as MOVED with a member regression because OTHER members of the merged TU fail no_mangled_refs: a reconstructed TU must spell the vague-linkage symbols _ZN7fBase_cnwEj, _ZN8dActor_cC2Ev and _ZN8dActor_cD2Ev directly or the 0x0211f0a4..0x02120668 range will not link, and the 24 still-unnamed func_ov071_* members reach fields through this-relative arithmetic. Byte-match outranks readability: 35/35 MATCH, objisolate clean, reloc destinations clean, emission order ROM-ascending, linkcheck SCRATCH-LINK-VERIFIED with tuRange differingBytes 0, and rombuild 106/106 modules exact against the stock ROM."}
 {"path": "src/_ZN10Scuttlebug6RenderEv.cpp", "reason": "ov071/Scuttlebug TU promotion, stage 3b: these five rows are a path rename, not a regression. Each of the five whole-file identities was absorbed verbatim into src/actors/Scuttlebug.cpp by the promotion and re-enters the set as a member identity (src/actors/Scuttlebug.cpp#Scuttlebug_SetState, #_ZN10Scuttlebug13OnYoshiTryEatEv, #_ZN10Scuttlebug16OnAimedAtWithEggEv, #_ZN10Scuttlebug16OnPendingDestroyEv, #_ZN10Scuttlebug6RenderEv), so the banked count is unchanged at 2686 -- 5 gained, 5 lost, same five units of code. classify_missing() reports them as MOVED with a member regression because OTHER members of the merged TU fail no_mangled_refs: a reconstructed TU must spell the vague-linkage symbols _ZN7fBase_cnwEj, _ZN8dActor_cC2Ev and _ZN8dActor_cD2Ev directly or the 0x0211f0a4..0x02120668 range will not link, and the 24 still-unnamed func_ov071_* members reach fields through this-relative arithmetic. Byte-match outranks readability: 35/35 MATCH, objisolate clean, reloc destinations clean, emission order ROM-ascending, linkcheck SCRATCH-LINK-VERIFIED with tuRange differingBytes 0, and rombuild 106/106 modules exact against the stock ROM."}
+{"path": "src/_ZN13OneUpMushroom13OnTurnIntoEggER6Player.cpp", "reason": "TU promotion ov002/da1up_c: 36 shards absorbed into src/actors/da1up_c.cpp, 36/36 MATCH and SCRATCH-LINK-VERIFIED (ROM sha256 identical to the baseline control). A reconstructed TU must spell vague-linkage symbols directly (_ZN7fBase_cnwEj, _ZN12dEnemyBase_cD2Ev) and must keep the 27 members whose ROM names are auto-generated func_ov002_* spelled that way, so no_mangled_refs and real_function_name cannot pass for the merged file. Byte match outranks readability."}
+{"path": "src/_ZN13OneUpMushroom13OnYoshiTryEatEv.cpp", "reason": "TU promotion ov002/da1up_c: 36 shards absorbed into src/actors/da1up_c.cpp, 36/36 MATCH and SCRATCH-LINK-VERIFIED (ROM sha256 identical to the baseline control). A reconstructed TU must spell vague-linkage symbols directly (_ZN7fBase_cnwEj, _ZN12dEnemyBase_cD2Ev) and must keep the 27 members whose ROM names are auto-generated func_ov002_* spelled that way, so no_mangled_refs and real_function_name cannot pass for the merged file. Byte match outranks readability."}
+{"path": "src/_ZN13OneUpMushroom16OnPendingDestroyEv.cpp", "reason": "TU promotion ov002/da1up_c: 36 shards absorbed into src/actors/da1up_c.cpp, 36/36 MATCH and SCRATCH-LINK-VERIFIED (ROM sha256 identical to the baseline control). A reconstructed TU must spell vague-linkage symbols directly (_ZN7fBase_cnwEj, _ZN12dEnemyBase_cD2Ev) and must keep the 27 members whose ROM names are auto-generated func_ov002_* spelled that way, so no_mangled_refs and real_function_name cannot pass for the merged file. Byte match outranks readability."}
+{"path": "src/_ZN13OneUpMushroomD0Ev.cpp", "reason": "TU promotion ov002/da1up_c: 36 shards absorbed into src/actors/da1up_c.cpp, 36/36 MATCH and SCRATCH-LINK-VERIFIED (ROM sha256 identical to the baseline control). A reconstructed TU must spell vague-linkage symbols directly (_ZN7fBase_cnwEj, _ZN12dEnemyBase_cD2Ev) and must keep the 27 members whose ROM names are auto-generated func_ov002_* spelled that way, so no_mangled_refs and real_function_name cannot pass for the merged file. Byte match outranks readability."}
+{"path": "src/_ZN13OneUpMushroomD1Ev.cpp", "reason": "TU promotion ov002/da1up_c: 36 shards absorbed into src/actors/da1up_c.cpp, 36/36 MATCH and SCRATCH-LINK-VERIFIED (ROM sha256 identical to the baseline control). A reconstructed TU must spell vague-linkage symbols directly (_ZN7fBase_cnwEj, _ZN12dEnemyBase_cD2Ev) and must keep the 27 members whose ROM names are auto-generated func_ov002_* spelled that way, so no_mangled_refs and real_function_name cannot pass for the merged file. Byte match outranks readability."}
