+++ config/converted-baseline.json
@@ -618,11 +618,6 @@
     "src/_ZN13MontyMoleRock6RenderEv.cpp",
     "src/_ZN13MontyMoleRockD0Ev.cpp",
     "src/_ZN13MontyMoleRockD1Ev.cpp",
-    "src/_ZN13OneUpMushroom13OnTurnIntoEggER6Player.cpp",
-    "src/_ZN13OneUpMushroom13OnYoshiTryEatEv.cpp",
-    "src/_ZN13OneUpMushroom16OnPendingDestroyEv.cpp",
-    "src/_ZN13OneUpMushroomD0Ev.cpp",
-    "src/_ZN13OneUpMushroomD1Ev.cpp",
     "src/_ZN13PeachPainting13InitResourcesEv.cpp",
     "src/_ZN13PeachPainting16CleanupResourcesEv.cpp",
     "src/_ZN13PeachPainting20UpdateModelTransformEv.cpp",
@@ -2156,6 +2151,11 @@
     "src/actors/dScMgSound_c.cpp#_ZN12dScMgSound_cD1Ev",
     "src/actors/dScMgTeresa_c.cpp#_ZN13dScMgTeresa_cD0Ev",
     "src/actors/dScMgTeresa_c.cpp#_ZN13dScMgTeresa_cD1Ev",
+    "src/actors/da1up_c.cpp#_ZN7da1up_c13OnTurnIntoEggER6Player",
+    "src/actors/da1up_c.cpp#_ZN7da1up_c13OnYoshiTryEatEv",
+    "src/actors/da1up_c.cpp#_ZN7da1up_c16OnPendingDestroyEv",
+    "src/actors/da1up_c.cpp#_ZN7da1up_cD0Ev",
+    "src/actors/da1up_c.cpp#_ZN7da1up_cD1Ev",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_c16CleanupResourcesEv",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_c6RenderEv",
     "src/actors/daDgr_c.cpp#_ZN7daDgr_cD0Ev",
