+++ include/decl_common.h
@@ -507,7 +507,6 @@ extern int _ZTV13FortressTower[];
 extern int _ZTV10daCamTag_c[];
 extern int _ZTV13KoopaTheQuick[];
 extern int _ZTV13MontyMoleRock[];
-extern int _ZTV13OneUpMushroom[];
 extern int _ZTV13PeachPainting[];
 extern int _ZTV14KnockDownPlank[];
 extern int _ZTV13PrincessPeach[];
