+++ notes/cpp-tu-current-state.md
@@ -21,7 +21,7 @@ python tools/cpp_tu_state.py --check-note
 | MATCHED functions | 11310 / 11392 (99.28%) |
 | MATCHED code bytes | 2184092 / 2238108 (97.59%) |
 | Strict CONVERTED source functions | 2696 / 11357 (23.74%) |
-| Physical production source files | 9334 |
+| Physical production source files | 9299 |
 
 MATCHED is recomputed from committed config symbols and `src/` through
 `progress.synced_from_src()`; no ambient or gitignored chaos database is accepted.
@@ -32,9 +32,9 @@ weighted by enrolled function ownership so TU consolidation does not move it.
 
 | Measure | Live value |
 | --- | ---: |
-| Tracked production source files | 9334 |
-| Tracked `.c` files | 5333 |
-| Tracked `.cpp` files | 4001 |
+| Tracked production source files | 9299 |
+| Tracked `.c` files | 5313 |
+| Tracked `.cpp` files | 3986 |
 | `.cpp` files missing first-line `//cpp` | 0 |
 | Mangled-symbol source files | 3949 |
 | Genuinely migrated C++ symbol files | 3869 |
@@ -44,9 +44,9 @@ weighted by enrolled function ownership so TU consolidation does not move it.
 | Delinks path-owned function-symbol records | 11273 |
 | Path-owned records still supplied from ROM bytes | 78 |
 | Source-built (`complete`) function-symbol records | 11195 |
-| Source-built (`complete`) source files | 9167 |
-| Complete sources owning more than one function | 124 |
-| Source-built records inside multi-function sources | 2152 |
+| Source-built (`complete`) source files | 9132 |
+| Complete sources owning more than one function | 125 |
+| Source-built records inside multi-function sources | 2188 |
 | Largest function-symbol-records-per-source count | 301 |
 | Complete source paths missing from git | 0 |
 
@@ -56,7 +56,7 @@ require the delinks entry's `complete` marker; entries without it are ROM-byte
 placeholders and are not reported as decompiled. The join remains correct when a
 source owns two functions; filenames are not treated as the authority.
 
-Current complete multi-function owners: `src/actors/ActorBase_SceneNode.cpp` (2 function-symbol records), `src/actors/ActorDerived.cpp` (5 function-symbol records), `src/actors/Eyerok.cpp` (34 function-symbol records), `src/actors/Goomboss.cpp` (36 function-symbol records), `src/actors/Player.cpp` (301 function-symbol records), `src/actors/Scuttlebug.cpp` (35 function-symbol records), `src/actors/dBgActor_c.cpp` (11 function-symbol records), `src/actors/dMgJump3DMario_c.cpp` (28 function-symbol records), `src/actors/dScGameOver_c.cpp` (9 function-symbol records), `src/actors/dScMgBomroom_c.cpp` (41 function-symbol records), and 114 more in `--json`.
+Current complete multi-function owners: `src/actors/ActorBase_SceneNode.cpp` (2 function-symbol records), `src/actors/ActorDerived.cpp` (5 function-symbol records), `src/actors/Eyerok.cpp` (34 function-symbol records), `src/actors/Goomboss.cpp` (36 function-symbol records), `src/actors/Player.cpp` (301 function-symbol records), `src/actors/Scuttlebug.cpp` (35 function-symbol records), `src/actors/dBgActor_c.cpp` (11 function-symbol records), `src/actors/dMgJump3DMario_c.cpp` (28 function-symbol records), `src/actors/dScGameOver_c.cpp` (9 function-symbol records), `src/actors/dScMgBomroom_c.cpp` (41 function-symbol records), and 115 more in `--json`.
 
 A `.cpp` extension alone proves nothing about compiler language mode; this build
 selects C++ only when the first line is exactly `//cpp`. The marker-missing count
@@ -72,16 +72,16 @@ partition to add to the migrated and unmigrated rows.
 | Measure | Live value |
 | --- | ---: |
 | Tracked `src_tu/` source files | 30 |
-| Manifest entries | 153 |
-| Functions named by manifest entries | 2778 |
-| Unique functions named by the manifest | 2767 |
+| Manifest entries | 154 |
+| Functions named by manifest entries | 2814 |
+| Unique functions named by the manifest | 2803 |
 | Modules represented | 43 |
-| Manifest shadow sources present in git | 153 |
+| Manifest shadow sources present in git | 154 |
 | Entries licensing non-text sections | 17 |
-| Entries actually production-enrolled at `promoted_source` | 123 |
+| Entries actually production-enrolled at `promoted_source` | 124 |
 | Existing promotion paths that disagree with delinks | 0 |
 
-Manifest statuses: `link-verified` 3, `promoted` 123, `text-verified` 27.
+Manifest statuses: `link-verified` 3, `promoted` 124, `text-verified` 27.
 
 `config/tu_manifest.d/` records reconstruction evidence and licensed ranges. It does
 not enroll a TU. The production number above counts an entry as promoted only when every
