--- src/_ZN13OneUpMushroom13OnYoshiTryEatEv.cpp
@@ -1,7 +1,0 @@
-//cpp
-// @symbol _ZN13OneUpMushroom13OnYoshiTryEatEv
-#include "OneUpMushroom.h"
-/* recovered: renamed to Class_Method */
-s32 OneUpMushroom::OnYoshiTryEat() {
-    return 4;
-}
