--- src/_ZN13OneUpMushroom16OnPendingDestroyEv.cpp
@@ -1,8 +1,0 @@
-//cpp
-// @symbol _ZN13OneUpMushroom16OnPendingDestroyEv
-
-#include "OneUpMushroom.h"
-
-void OneUpMushroom::OnPendingDestroy()
-{
-}
