+++ src/d_a_1up_scaleup_kinoko.c
@@ -8,7 +8,7 @@
 #include "decl_dBgCh_Actr.h"
 #include "decl_common.h"
 /* recovered: vtable identified, globals resolved */
-/* resolved: VT0 = _ZTV13OneUpMushroom */
+/* resolved: VT0 = _ZTV7da1up_c */
 /* Reconstructed source-style name: SM64DS proves da1up_c through RTTI,
  * allocation size, vtable identity, and the SCALEUP_KINOKO registry profile;
  * later EAD lineage supplies classInit. Exact original spelling is not
@@ -18,7 +18,7 @@ int *da1up_c_classInit_SCALEUP_KINOKO(void)
     int *p = (int *)_ZN7fBase_cnwEj(920);
     if (p) {
         _ZN12dEnemyBase_cC2Ev(p);
-        p[0] = (int)_ZTV13OneUpMushroom;
+        p[0] = (int)_ZTV7da1up_c;
         _ZN7dCcAc_cC1Ev((char *)p + 0x110);
         _ZN10dBgCh_ActrC1Ev((char *)p + 0x144);
         _ZN5ModelC1Ev((char *)p + 0x300);
