--- src/func_ov002_020aefa4.c
@@ -1,4 +1,0 @@
-void func_ov002_020aefa4(char *self)
-{
-    *(unsigned int *)(self + 0x12c) |= 0x8000;
-}
