--- src/func_ov002_020af908.c
@@ -1,6 +1,0 @@
-extern void func_ov002_020af924(char *self);
-
-void func_ov002_020af908(char *self) {
-    *(short *)(int)(self + 0x8e) += 0xc00;
-    func_ov002_020af924(self);
-}
