--- src/func_ov002_020afa50.c
@@ -1,11 +1,0 @@
-// @symbol func_ov002_020afa50
-// recovered name: dCapEnemy_c_Kill
-/* recovered: renamed to Class_Method, declarations from a shared header */
-#include "decl_common.h"
-/* recovered: renamed to Class_Method */
-/* dCapEnemy_c::Kill - recovered from vtable slot identity */
-
-void func_ov002_020afa50(char *self) {
-    *(short *)(int)(self + 0x8e) += 0xc00;
-    func_ov002_020afa6c(self);
-}
