--- src/func_ov002_020afa6c.c
@@ -1,8 +1,0 @@
-extern int func_ov002_020af1dc(void*);
-extern int func_ov002_020af684(void*, int, int);
-int func_ov002_020afa6c(char* c){
-  int r=func_ov002_020af1dc(c);
-  if(!r) return r;
-  r=func_ov002_020af684(c, 5, r);
-  return r;
-}
