--- src/func_ov002_020afc44.c
@@ -1,6 +1,0 @@
-extern int func_ov002_020af3a8(void*);
-extern int func_ov002_020af218(void*, int);
-int func_ov002_020afc44(char* c){
-  func_ov002_020af3a8(c);
-  return func_ov002_020af218(c, 0xbb8);
-}
